Updated Data repostiory get more accurate travel details
Old: Closest grocery store was calculated by nearby search api response which returned the grocery store with the lowest distance. this did not take into account roads or anything. That grocery store was the used in a distance matrix calculation which returned travel times.

Now instead of only using the closest grovery store from the places api, it used the two closest. It then determines which is the actual closest store based on the travel times rather than soley on distance